src/requirements.txt
mysql-connector-python==8.0.33
PyQt5==5.15.9
numpy==1.24.1